Completed revision of bibliography
Refs #1

Includes some new papers, some that had not been included in the previous version. This version is for a longer paper, it has got to be more inclusive and exhaustive. One of the papers included, for instance, has two fuzzy controllers for steering and accelerating, although it's optimized for imitation, not winning games.